+++ src/awsData.ts
@@ -120,6 +120,7 @@ type RefOverrides = {
     "AWS::StackId": string,
     "AWS::StackName": string,
     "AWS::URLSuffix": string,
+    "AWS::Partition": string,
     [refOverride: string]: any
 };