+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 - Merge PR #177, fix nested resources within an if condition not being resolved. Fix resolution of AWS::NoValue.
+- Merge PR #179, fixing support for AWS::Partition pseudo parameter
 
 ## [1.7.3] - 2018-06-30
 ### Fixed